MNT: consolidate exceptions and some relative imports
